langgraph-hybrid-rag/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
langgraph==0.2.34
langchain-openai==0.1.23
openai==1.43.0
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
pypdf==4.3.1
python-docx==1.1.2
numpy==1.26.4
scikit-learn==1.5.1
requests==2.32.3

httpx==0.27.2
